Rename command to version update
